attempt to do lqr with 0 adhesion
Attempt to do lqr with adhesion in a way that returns exactly the same
result as lqr with adjesion actuator not included in the model.
Seemingly get_ctrl0 generates a slightly different result, so this may
not be viable. Further commits will rollback some of this code: instead
of deleting indices corresponding to adhesion, the zero result of this
will be carried through.